+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## x.x.x
 * Add back Terms flow.
-* Add `title`, `advertiser`, `body`, and `callToAction` to ad object of native ad UI callbacks.
+* Add `title`, `advertiser`, `body`, `callToAction`, `isIconImageAvailable`, `isOptionsViewAvailable`, and `isMediaViewAvailable`  to ad object of native ad UI callbacks.
 * Fix a warning of `image source of null` with `AppLovinMAX.NativeAdView.IconView`.
 ## 5.0.2
 * Gracefully handle inter and rewarded ad calls when current `Activity` is null.